port existing code
The fileguard module was originally developed as part of the testing tooling
for a different project. In the early stages of its development I realized
that it would make more sense for it to be a separate project, since it's
something useful for future projects I might have.

With this commit, I basically ported the existing codebase with a few
changes, such as names.